pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "web2spec"
version = "0.1.0"
description = "Proof-of-concept pipeline for crawling websites into structured product documentation."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "anthropic>=0.40.0",
  "langgraph>=0.2.0",
  "openai>=1.54.0",
  "Pillow>=10.4.0",
  "playwright>=1.48.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
]

[project.scripts]
web2spec = "web2spec.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

